Move shotwrapper in dedicated feature so we'll try to disable Ammunition
